+++ common/src/main/java/com/xtracr/realcamera/api/VirtualRenderer.java
@@ -3,10 +3,10 @@ package com.xtracr.realcamera.api;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.function.BiPredicate;
-import java.util.function.Supplier;
 
-import com.xtracr.realcamera.command.ClientCommand;
+import com.xtracr.realcamera.RealCameraCore;
 import com.xtracr.realcamera.config.ConfigFile;
+import com.xtracr.realcamera.config.ModConfig;
 
 import net.minecraft.client.util.math.MatrixStack;
 
@@ -17,22 +17,17 @@ public class VirtualRenderer {
 
     private static final Map<String, BiPredicate<Float, MatrixStack>> functionProvider = new HashMap<>();
 
-    public static void register(String modid, BiPredicate<Float, MatrixStack> function) {
-        functionProvider.put(modid, function);
-    }
+    public static final ModConfig config = ConfigFile.modConfig;
 
     /**
      * 
      * @param modid    {@code mandatory}
      * @param function {@code mandatory} turn to vanilla rendering if return true.
      *                 {@link CompatExample#virtualRender See example here}
-     * @param feedback {@code optional}
-     *                 sent when command {@code \realcamera feedback} is executed
      * 
      */
-    public static void register(String modid, BiPredicate<Float, MatrixStack> function, Supplier<String> feedback) {
+    public static void register(String modid, BiPredicate<Float, MatrixStack> function) {
         functionProvider.put(modid, function);
-        ClientCommand.registerFeedback(() -> "[" + modid + "]: " + feedback.get());
     }
 
     /**
@@ -42,11 +37,15 @@ public class VirtualRenderer {
      * 
      */
     public static String getModelPartName() {
-        return ConfigFile.modConfig.getModModelPartName();
+        return config.getModModelPartName();
     }
 
+    public static boolean shouldDisableRender(String modelPartName) {
+        return RealCameraCore.isRenderingWorld && config.shouldDisableRender(modelPartName) && RealCameraCore.isActive();
+    }
+    
     public static boolean virtualRender(float tickDelta, MatrixStack matrixStack) {
-        return functionProvider.get(ConfigFile.modConfig.getModelModID()).test(tickDelta, matrixStack);
+        return functionProvider.get(config.getModelModID()).test(tickDelta, matrixStack);
     }
 
     public static String[] getModidList() {