+++ fabric/src/main/java/com/xtracr/realcamera/RealCameraFabric.java
@@ -17,6 +17,7 @@ public class RealCameraFabric implements ClientModInitializer {
 
         ClientTickEvents.END_CLIENT_TICK.register(KeyBindings::handle);
         WorldRenderEvents.START.register(EventHandler::onWorldRenderStart);
+        WorldRenderEvents.END.register(context -> RealCameraCore.isRenderingWorld = false);
 
         KeyBindingHelper.registerKeyBinding(KeyBindings.TOGGLE_PERSPECTIVE);
         KeyBindingHelper.registerKeyBinding(KeyBindings.TOGGLE_ADJUST_MODE);