+++ forge/src/main/java/com/xtracr/realcamera/EventHandler.java
@@ -1,10 +1,8 @@
 package com.xtracr.realcamera;
 
-import com.xtracr.realcamera.command.ClientCommandForge;
 import com.xtracr.realcamera.config.ConfigFile;
 import com.xtracr.realcamera.utils.CrosshairUtils;
 
-import net.minecraftforge.client.event.RegisterClientCommandsEvent;
 import net.minecraftforge.client.event.RenderLevelStageEvent;
 import net.minecraftforge.client.event.InputEvent.KeyInputEvent;
 import net.minecraft.client.MinecraftClient;
@@ -31,16 +29,15 @@ public class EventHandler {
     }
 
     @SubscribeEvent
-    public static void onClientCommandRegister(RegisterClientCommandsEvent event) {
-        ClientCommandForge.INSTANCE.register(event.getDispatcher());
-    }
-
-    @SubscribeEvent
     public static void onRenderWorldStage(RenderLevelStageEvent event) {
-        if (RenderLevelStageEvent.Stage.AFTER_SKY.equals(event.getStage()) && ConfigFile.modConfig.isCrosshairDynamic()
-                && RealCameraCore.isActive()) {
-            CrosshairUtils.update(MinecraftClient.getInstance(), event.getCamera(),
+        if (RenderLevelStageEvent.Stage.AFTER_SKY.equals(event.getStage())) {
+            RealCameraCore.isRenderingWorld = true;
+            if (ConfigFile.modConfig.isCrosshairDynamic() && RealCameraCore.isActive()) {
+                CrosshairUtils.update(MinecraftClient.getInstance(), event.getCamera(),
                     event.getPoseStack().peek().getPositionMatrix(), event.getProjectionMatrix());
+            }
+        } else if (RenderLevelStageEvent.Stage.AFTER_WEATHER.equals(event.getStage())) {
+            RealCameraCore.isRenderingWorld = false;
         }
     }