+++ forge/src/main/java/com/xtracr/realcamera/RealCameraForge.java
@@ -30,7 +30,6 @@ public class RealCameraForge {
 
         MinecraftForge.EVENT_BUS.addListener(EventHandler::onKeyInput);
         MinecraftForge.EVENT_BUS.addListener(EventHandler::onCameraUpdate);
-        MinecraftForge.EVENT_BUS.addListener(EventHandler::onClientCommandRegister);
         MinecraftForge.EVENT_BUS.addListener(EventHandler::onRenderWorldStage);
 
         if (ModList.get().isLoaded("cloth_config")) {